Refactoring methods
